pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qa-intelli-media-comparator"
version = "1.0.0"
description = "AI-based media comparison microservice for Smartphone Camera QA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "opencv-python>=4.10.0.84",
    "pillow>=10.4.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-image>=0.24.0",
    "pyiqa>=0.1.13",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "rich>=13.8.1",
    "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[project.scripts]
qimc = "qa_intelli_media_comparator.main:cli_app"

[tool.setuptools.packages.find]
where = ["."]
include = ["qa_intelli_media_comparator*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=qa_intelli_media_comparator --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*"]
